requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
numpy==2.2.3
pandas==2.2.3
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
pytz==2025.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
